# [IMP] sale: improve performance of _compute_expected_date for large SOs