@robodoo 2024-11-14
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/187186.png)](https://mergebot.odoo.com/odoo/odoo/pull/187186)