@emi-odoo 2024-11-14
The following screenshots are taken from speedscope of `search_read` produced from the list view. The `customer_lead` has been randomized in the sale.order.line to get the worst case scenario

Before:
![image](https://github.com/user-attachments/assets/6f0691e0-43cd-413c-ad44-be467737a21c)
Before (left heavy):
![image](https://github.com/user-attachments/assets/9b78dc0d-69e3-4e60-a255-52404333ea93)

After:
![image](https://github.com/user-attachments/assets/f1fa815c-8887-4f0d-bd5f-c88c24c88b43)
After (left heavy):
![image](https://github.com/user-attachments/assets/44d64a97-8da6-467c-97e0-bc7280f3de8e)

Best case scenario (customer_lead is always 0)

Before:
![image](https://github.com/user-attachments/assets/d2208541-aeb0-4bf4-a15e-e07a4ffc5005)

After:
![image](https://github.com/user-attachments/assets/dfbe1e66-a7d9-4694-ac36-fc216f96a3e2)


At worst: 1s vs 1.4s (~0.4 seconds faster)
At best: 36.95 ms vs 1.36s (~1 second faster)